path: add an entry to search up to the current directory
By adding the ./**;,, entry, commands such as 'gf' will search
upwards up until the current directory. The Rooter plugin
will set the current directory to the root of the git repo
so this will basically search the entire project.
This makes it possible, for example, to open included templates
in a Zend Framework directory, which are relative to the
templates directory instead of the project-root directory