sessions: don't save curdir
When curdir is not saved, vim will write full paths to the session file,
instead of relative paths to the curdir. This is needed because
autochdir changes the current dir while reloading a session, thus
causing vim not to find files when they have a relative path. The other
solution would be to disable autochdir.